requirements.txt
scapy==2.5.0
pycryptodome==3.19.0
Pillow==9.0.1
matplotlib